pyproject.toml
[project]
name = "deltallm"
version = "0.9.0"
description = "Open-source LLM gateway for developers. One API, 8+ providers, enterprise-ready."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "DeltaLLM Contributors", email = "everythingjson@gmail.com"}
]
keywords = ["llm", "openai", "anthropic", "gateway", "proxy", "litellm", "ai", "claude", "gpt", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    # Core
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
    
    # Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0.0",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    
    # Cache
    "redis>=5.0.0",
    
    # Token counting
    "tiktoken>=0.6.0",
    "tokenizers>=0.15.0",
    
    # Config
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    
    # Validation
    "email-validator>=2.1.0",
    
    # Utils
    "orjson>=3.9.0",
    "shortuuid>=1.0.11",
    "pytz>=2024.1",

    # Security
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",
    "aioresponses>=0.7.6",
    "ruff>=0.2.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "locust>=2.20.0",
]

providers = [
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "cohere>=5.0.0",
    "google-generativeai>=0.3.0",
    "boto3>=1.34.0",
    "mistralai>=0.1.0",
    "groq>=0.4.0",
]

all = [
    "deltallm[dev,providers]"
]

[project.scripts]
deltallm = "deltallm.proxy.app:cli"

[project.urls]
Homepage = "https://github.com/mehditantaoui/deltallm"
Documentation = "https://deltallm.readthedocs.io"
Repository = "https://github.com/mehditantaoui/deltallm"
Issues = "https://github.com/mehditantaoui/deltallm/issues"
Changelog = "https://github.com/mehditantaoui/deltallm/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["deltallm"]

[tool.hatch.build.targets.sdist]
include = [
    "/deltallm",
    "/tests",
    "/config",
    "/docs",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
follow_imports = "silent"
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=deltallm",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["deltallm"]
branch = true
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/ui/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[dependency-groups]
dev = [
    "psycopg2-binary>=2.9.11",
]
